Update yt-dlpan.sh
Added a feature to findremoved. Prevents the function from running if it was last run on that folder very recently.
This is necessary since this is features makes too many requests and consumes a lot of time. I had tested the statement and ensured it works, but I can't use a real scenario as a test right now.