Update yt-dlpan.sh
findremoved now creates a copy in the case that you have the thumbnail or other info related to the video, but not the video itself. Previously, if tracking was enabled, comm prevented this from happening because it excluded lost media that was not video. With tracking disabled, you need both archives to be used to determine all the lost media that you had downloaded, not just the video archive

With tracking enabled. You should not use the thumbnail archive for reasons that are too long to read and I probably explained before.