Update yt-dlpan.sh
"Will not work for a group of archives" then run it for each archives. Don't use the recurse directories feature I added as a replacement.

To clarify. I say it doesn't work because it never gets added to the archive when you download outside the function, causing it to always redownload and revert changes. It's this function that adds it to the archive, but you can't give yt-dlp multiple archive files so it won't work there.